# Conflicts:
#	chart/src/commonMain/kotlin/com/aay/compose/baseComponents/XAxisDrawing.kt
#	chart/src/commonMain/kotlin/com/aay/compose/baseComponents/grid.kt
#	chart/src/commonMain/kotlin/com/aay/compose/lineChart/ChartContent.kt
#	chart/src/commonMain/kotlin/com/aay/compose/lineChart/LineChart.kt
#	chart/src/commonMain/kotlin/com/aay/compose/lineChart/components/DefaultLines.kt
#	chart/src/commonMain/kotlin/com/aay/compose/lineChart/components/QuadraticLines.kt
#	common/src/commonMain/kotlin/com/aay/common/LineChartSample.kt

+++ chart/src/commonMain/kotlin/com/aay/compose/pieChart/model/PieChartData.kt
@@ -3,7 +3,7 @@ package com.aay.compose.pieChart.model
 import androidx.compose.ui.graphics.Color
 
 data class PieChartData(
-    val data:Float,
+    val data:Double,
     val color: Color,
     val partName:String
 )
